Estilo

+++ surefugio-app/src/app/animales/componentes/mis-animales/mis-animales.component.ts
@@ -33,8 +33,10 @@ export class MisAnimalesComponent implements OnInit {
 
   deleteAnimal(animal: Animal){
     console.log(animal);
-
-    this.store.dispatch(deleteAnimal({animal}));
+    if (confirm('¿Seguro que quieres eliminar este animal?')) {
+      this.store.dispatch(deleteAnimal({animal}));
+    }
+    
 
   }
